requirements.txt
praw
ollama
python-dotenv
tqdm